+++ CHANGELOG.md
@@ -5,11 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [4.0.3] - 2023-09-27
 ### Changed
 - ID type is now optional for docV
 
 ### Added
-- Added Enhanced Document Verification to the SmileIdentityCore class.
+- Enhanced Document Verification to the SmileIdentityCore class.
 - Business verification to the SmileIdentityCore class.
 - JobType, ImageType, and BusinessVerificationType enum classes
 - [examples](/examples) folder